Add ability to use Observables with async/await
A lot of modern JavaScript code uses async/await to handle asynchronous code. This is a great way to write code that is easy to read and understand. However, it is not possible to use async/await with Observables outside of the `forEach()` which requries more boilerplate than needed in order to do so.

Changes:
 * Adds `.all()` to an observable instance allowing one to `await` the values it contains
 * Adds a test for the new functionality